client/requirements.txt
fuse-python==1.0.8
